backend/requirements.txt
fastapi[all]
pydantic
faster-whisper
git+https://github.com/huggingface/parler-tts.git
git+https://github.com/myshell-ai/MeloTTS.git

deepfilternet
pymongo[srv]

tortoise-tts